-->itoa

+++ ft_bzero.c
@@ -1,6 +1,19 @@
-#include "libft.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_bzero.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ibaby <ibaby@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/31 22:33:26 by ibaby             #+#    #+#             */
+/*   Updated: 2024/03/31 22:33:26 by ibaby            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-void    ft_bzero(void *array, int lim)
+#include "libft.h"
+#include <stddef.h>
+
+void    ft_bzero(void *array, size_t lim)
 {
     ft_memset(array, 0, lim);
 }
