+++ ft_memcpy.c
@@ -1,6 +1,19 @@
-#include "libft.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memcpy.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ibaby <ibaby@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/31 22:34:09 by ibaby             #+#    #+#             */
+/*   Updated: 2024/03/31 22:34:09 by ibaby            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-void    *ft_memcpy(void *dest, const void *src, int n)
+#include "libft.h"
+#include <stddef.h>
+
+void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
     int                 i;
     unsigned const char *s;
