+++ ft_memset.c
@@ -1,6 +1,19 @@
-#include "libft.h"
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_memset.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ibaby <ibaby@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/03/31 22:34:14 by ibaby             #+#    #+#             */
+/*   Updated: 2024/03/31 22:34:14 by ibaby            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
-void    *ft_memset(void *array, int value, int hmany)
+#include "libft.h"
+#include <stddef.h>
+
+void    *ft_memset(void *array, int value, size_t hmany)
 {
     int     i;
     unsigned char   *tab;
